Execute clean only if needed

+++ Sources/Cache/CacheStore.swift
@@ -8,9 +8,10 @@
 import Foundation
 
 public class CacheStore {
+    let asyncQueue: DispatchQueue
     private let diskSetting: DiskSetting
     private let fileManager: FileManager
-    let asyncQueue: DispatchQueue
+    private var nextCleanupExpiredFileNeeded = Date()
     
     private typealias File = (date: Date, size: Int, filePath: String)
     
@@ -21,7 +22,7 @@ public class CacheStore {
     }
     
     public func cleanup() throws {
-        deleteExpiredFiles()
+        deleteExpiredFilesIfNeeded()
         if size() > diskSetting.maxSize.byte() {
             let leftPaths = try self.paths() ?? []
             var sorted: [File] = []
@@ -49,13 +50,13 @@ public class CacheStore {
     }
     
     public func delete(name: String) throws {
-        deleteExpiredFiles()
+        deleteExpiredFilesIfNeeded()
         guard let path = locationPath() else { throw CacheError.invalidFilePath }
         try fileManager.removeItem(atPath: path.appendingPathComponent(name.toBase64()).relativePath)
     }
     
     public func load<T: Cachable>(name: String, type: T.Type) throws -> T {
-        deleteExpiredFiles()
+        deleteExpiredFilesIfNeeded()
         guard let path = locationPath() else { throw CacheError.invalidFilePath }
         let data = fileManager.contents(atPath: path.appendingPathComponent(name.toBase64()).relativePath)
         return T(name: name, data: data)
@@ -74,7 +75,7 @@ public class CacheStore {
     }
     
     public func info(name: String) throws -> FileInformation {
-        deleteExpiredFiles()
+        deleteExpiredFilesIfNeeded()
         guard let path = locationPath()?.appendingPathComponent(name.toBase64()).relativePath else { throw CacheError.invalidFilePath }
         let infos = try fileManager.attributesOfItem(atPath: path)
         let size = (infos[FileAttributeKey.size] as? NSNumber)?.intValue ?? -1
@@ -128,22 +129,31 @@ public class CacheStore {
         }
     }
     
-    private func deleteExpiredFiles() {
-        DispatchQueue.global(qos: .utility).async {
+    private func deleteExpiredFilesIfNeeded() {
+        if self.nextCleanupExpiredFileNeeded > Date() {
+            return // Nothing to clean
+        }
+        DispatchQueue.global(qos: .utility).async { [weak self] in
+            guard let self = self else { return }
             do {
-            let names = try self.paths() ?? []
-                guard let path = self.locationPath() else {
-                throw CacheError.invalidFilePath
-            }
-            let currentDate = Date()
-            for name in names {
-                let filePath = path.appendingPathComponent(name).relativePath
-                guard let date = try self.fileManager.attributesOfItem(atPath: filePath)[FileAttributeKey.creationDate] as? Date else { continue }
-                if currentDate.timeIntervalSince(date) > self.diskSetting.storeDuration.timeInterval() {
-                    try self.fileManager.removeItem(atPath: filePath)
+                let names = try self.paths() ?? []
+                    guard let path = self.locationPath() else {
+                    throw CacheError.invalidFilePath
                 }
-            }
-            } catch {}
+                    
+                let currentDate = Date()
+                var oldestFileDate = currentDate
+                for name in names {
+                    let filePath = path.appendingPathComponent(name).relativePath
+                    guard let date = try self.fileManager.attributesOfItem(atPath: filePath)[FileAttributeKey.creationDate] as? Date else { continue }
+                    if currentDate.timeIntervalSince(date) > self.diskSetting.storeDuration.timeInterval() {
+                        try self.fileManager.removeItem(atPath: filePath)
+                    } else if oldestFileDate > date {
+                        oldestFileDate = date
+                    }
+                }
+                self.nextCleanupExpiredFileNeeded = oldestFileDate.addingTimeInterval(self.diskSetting.storeDuration.timeInterval())
+            } catch { }
         }
     }
 }
